[bluez] Make AVRCP version and features configurable. Contributes to JB#10051
In order to have better control over what is advertised over SDP and
what features are made available, added parsing of AVRCP configuration
entries.

TG and CT can be individually disabled; if TG version is specified
to be below 0x0103, media player registrations will be rejected and
the related functionality is essentially disabled.