[bluez] Initialize non-initialized adapter after startup. Fixes JB#10925.
There was a problem with bluetoothd starting up when BT was rfkilled;
hciops startup for the BT adapter(s) would not complete and the adapter
list would remain empty. This prevented powering adapters on when rfkill
was lifted.

Fixed by running the startup functionality again when rfkill is lifted
if the adapter is unknown (i.e. was not initialized at startup).