[bluez] Output only needed call status indicators. Fixes JB#11820.
In addition to not sending extra hold indicators, it is necessary
to suppress extraneous call status indicators just as well. Ofono
creates two status indications for e.g. when the active call is
released and the held/waiting call is activated; this would cause
bluetoothd to emit call=0 (for the released call) followed by a
call=1 (for the activated call). This causes problems, so output
also call status only when the state of both calls is known.