Reconnect when boths sides tried to connect. Fixes JB#13887.
In case an incoming connection was dropped when AVCTP connection
attempt was ongoing and the attempt failed, retry the connection
a few times. Otherwise no AVCTP connection might be formed at all;
there are headsets which disallow incoming connections when they
are connecting themselves, but won't retry after a failure.