[bluez] Always have an AVRCP player ready
Upstream code loses AVRCP PGUs like notification registrations if
there is no registered MediaPlayer at the time AVCTP connection is
made; there will not be a handler for those PDUs present. As a result,
headsets may not work as expected if there is no player running when
they connect.

Modified the code so that an AVRCP player is created on first AVCTP
connection, and it is not tied to any MediaPlayer. If there are no
MediaPlayers present, dummy answers will be given to requests;
otherwise requests are proxied to the first registered MediaPlayer.