[bluez] Emit name changes also when adapter is down. Fixes MER#899.
If an adapter is down when its name property is changed, there won't
be a HCI command response for local name change that triggers sending
a property change signal.

Changed the code to send the signal immediately when the new name
string is stored internally if the adapter is down.